+++ packages/mountainsort2/src/p_consolidate_clusters.cpp
@@ -15,7 +15,7 @@ bool should_use_template(const Mda32& template0, Consolidate_clusters_opts opts)
 bool p_consolidate_clusters(QString clips, QString labels_path, QString labels_out, Consolidate_clusters_opts opts)
 {
     QVector<int> labels = P_consolidate_clusters::read_labels(labels_path);
-    int L = labels.count();
+    bigint L = labels.count();
 
     int K = MLCompute::max(labels);
 
@@ -54,7 +54,7 @@ namespace P_consolidate_clusters {
 QVector<int> read_labels(QString path)
 {
     Mda X(path);
-    int L = X.totalSize();
+    bigint L = X.totalSize();
     QVector<int> ret(L);
     for (int i = 0; i < L; i++)
         ret[i] = X.value(i);
@@ -71,15 +71,15 @@ bool write_labels(QString path, const QVector<int>& labels)
 Mda32 compute_template(QString clips_path, const QVector<int>& labels, int k)
 {
     DiskReadMda32 clips(clips_path);
-    int M = clips.N1();
-    int T = clips.N2();
+    bigint M = clips.N1();
+    bigint T = clips.N2();
     Mda sum(M, T);
     double ct = 0;
-    for (int i = 0; i < labels.count(); i++) {
+    for (bigint i = 0; i < labels.count(); i++) {
         if (labels[i] == k) {
             Mda32 tmp;
             clips.readChunk(tmp, 0, 0, i, M, T, 1);
-            for (int j = 0; j < M * T; j++) {
+            for (bigint j = 0; j < M * T; j++) {
                 sum.set(sum.get(j) + tmp.get(j), j);
             }
             ct++;
@@ -87,7 +87,7 @@ Mda32 compute_template(QString clips_path, const QVector<int>& labels, int k)
     }
     Mda32 ret(M, T);
     if (ct) {
-        for (int j = 0; j < M * T; j++) {
+        for (bigint j = 0; j < M * T; j++) {
             ret.set(sum.get(j) / ct, j);
         }
     }
@@ -97,16 +97,16 @@ bool should_use_template(const Mda32& template0, Consolidate_clusters_opts opts)
 {
     int peak_location_tolerance = 10;
 
-    int M = template0.N1();
-    int T = template0.N2();
-    int Tmid = (int)((T + 1) / 2) - 1;
+    bigint M = template0.N1();
+    bigint T = template0.N2();
+    bigint Tmid = (int)((T + 1) / 2) - 1;
 
     if (opts.central_channel > 0) {
         QVector<double> energies(M);
         energies.fill(0);
         double max_energy = 0;
-        for (int t = 0; t < T; t++) {
-            for (int m = 0; m < M; m++) {
+        for (bigint t = 0; t < T; t++) {
+            for (bigint m = 0; m < M; m++) {
                 double val = template0.value(m, t);
                 energies[m] += val * val;
                 if ((m != opts.central_channel - 1) && (energies[m] > max_energy))
@@ -118,10 +118,10 @@ bool should_use_template(const Mda32& template0, Consolidate_clusters_opts opts)
     }
 
     double abs_peak_val = 0;
-    int abs_peak_ind = 0;
-    int abs_peak_chan = 0;
-    for (int t = 0; t < T; t++) {
-        for (int m = 0; m < M; m++) {
+    bigint abs_peak_ind = 0;
+    bigint abs_peak_chan = 0;
+    for (bigint t = 0; t < T; t++) {
+        for (bigint m = 0; m < M; m++) {
             double value = template0.value(m, t);
             if (fabs(value) > abs_peak_val) {
                 abs_peak_val = fabs(value);
