+++ packages/mountainsort2/src/p_extract_segment_timeseries.cpp
@@ -18,7 +18,7 @@ Mda32 extract_channels_from_chunk(const Mda32 &chunk,const QList<int> &channels)
 bool p_extract_segment_timeseries(QString timeseries, QString timeseries_out, bigint t1, bigint t2,const QList<int> &channels)
 {
     DiskReadMda32 X(timeseries);
-    int M = X.N1();
+    bigint M = X.N1();
     //bigint N=X.N2();
     bigint N2 = t2 - t1 + 1;
 
@@ -52,7 +52,7 @@ bool p_extract_segment_timeseries_from_concat_list(QStringList timeseries_list, 
 {
     QString ts0 = timeseries_list.value(0);
     DiskReadMda32 X(ts0);
-    int M = X.N1();
+    bigint M = X.N1();
     if (!channels.isEmpty()) {
         M=channels.count();
     }
@@ -60,22 +60,22 @@ bool p_extract_segment_timeseries_from_concat_list(QStringList timeseries_list, 
     QVector<bigint> start_timepoints, end_timepoints;
     start_timepoints << 0;
     end_timepoints << durations.value(0) - 1;
-    for (int i = 1; i < durations.count(); i++) {
+    for (bigint i = 1; i < durations.count(); i++) {
         start_timepoints << end_timepoints[i - 1] + 1;
         end_timepoints << start_timepoints[i] + durations[i] - 1;
     }
 
-    int ii1 = 0;
+    bigint ii1 = 0;
     while ((ii1 < timeseries_list.count()) && (t1 > end_timepoints[ii1])) {
         ii1++;
     }
-    int ii2 = ii1;
+    bigint ii2 = ii1;
     while ((ii2 + 1 < timeseries_list.count()) && (start_timepoints[ii2 + 1] <= t2)) {
         ii2++;
     }
 
     Mda32 out(M, t2 - t1 + 1);
-    for (int ii = ii1; ii <= ii2; ii++) {
+    for (bigint ii = ii1; ii <= ii2; ii++) {
         DiskReadMda32 Y(timeseries_list.value(ii));
         Mda32 chunk;
 
@@ -137,7 +137,7 @@ QVector<bigint> get_durations(QStringList timeseries_list)
 Mda32 extract_channels_from_chunk(const Mda32 &chunk,const QList<int> &channels) {
     Mda32 ret(channels.count(),chunk.N2());
     for (bigint i=0; i<chunk.N2(); i++) {
-        for (int j=0; j<channels.count(); j++) {
+        for (bigint j=0; j<channels.count(); j++) {
             ret.setValue(chunk.value(channels[j]-1,i),j,i);
         }
     }
@@ -148,16 +148,16 @@ Mda32 extract_channels_from_chunk(const Mda32 &chunk,const QList<int> &channels)
 bool p_extract_segment_firings(QString firings_path, QString firings_out_path, bigint t1, bigint t2)
 {
     Mda firings(firings_path);
-    QVector<int> inds;
-    for (int i = 0; i < firings.N2(); i++) {
+    QVector<bigint> inds;
+    for (bigint i = 0; i < firings.N2(); i++) {
         double t0 = firings.value(1, i);
         if ((t1 <= t0) && (t0 <= t2))
             inds << i;
     }
     Mda ret(firings.N1(), inds.count());
-    for (int i = 0; i < inds.count(); i++) {
-        int j = inds[i];
-        for (int r = 0; r < firings.N1(); r++) {
+    for (bigint i = 0; i < inds.count(); i++) {
+        bigint j = inds[i];
+        for (bigint r = 0; r < firings.N1(); r++) {
             ret.setValue(firings.value(r, j), r, i);
         }
         ret.setValue(ret.value(1, i) - t1, 1, i);
